[jarun#637] moved bukuserver API to a separate file
